perf: Data transfer from 48 to 32 bytes
Size is a set value. I don't see the need to change the size of the
boids so I decided to keep it a set size of 16.. for now. Can change if
needed later on.

Note to self: I don't think I'd need the Alpha channel of colors. It'll
be 1.0f most of the time I think. So I can use that for size float if
needed.